Qwen2.5-VL/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.6.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.16=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
attrs=25.3.0=pypi_0
av=14.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
datasets=3.5.0=pypi_0
decord=0.6.0=pypi_0
deepspeed=0.16.4=pypi_0
dill=0.3.8=pypi_0
einops=0.8.1=pypi_0
filelock=3.18.0=pypi_0
flash-attn=2.7.4.post1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
hjson=3.1.0=pypi_0
huggingface-hub=0.30.2=pypi_0
idna=3.10=pypi_0
jinja2=3.1.6=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markupsafe=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.4.3=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
ninja=1.11.1.4=pypi_0
numpy=2.2.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-ml-py=12.570.86=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.2.1=pypi_0
pip=25.0=py311h06a4308_0
propcache=0.3.1=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=19.0.1=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
qwen-vl-utils=0.0.10=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.3=pypi_0
setuptools=75.8.0=py311h06a4308_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.1=pypi_0
torch=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.50.0=pypi_0
triton=3.2.0=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
wheel=0.45.1=py311h06a4308_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.20.0=pypi_0
zlib=1.2.13=h5eee18b_1
